foc and mitsubishi encoder implemented

+++ Firmware/servo407/Core/Inc/parameter_set.h
@@ -27,7 +27,7 @@ typedef struct _parameter_set_t{
 	float motor_K;
 
 	motor_feedback_type_t motor_feedback_type;  //type of feedback
-	int16_t encoder_electric_angle_correction;  //correction of difference in electric angle calculated from encoder and electric angle of stator current. needed to achieve 90/-90 torque angle
+	float encoder_electric_angle_correction;  //correction of difference in electric angle calculated from encoder and electric angle of stator current. needed to achieve 90/-90 torque angle
 	uint16_t encoder_resolution;
 
 	float current_filter_ts; //measured Id/Iq-low pass filter time constant for torque regulator
